requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==4.2.2
certifi==2022.12.7
python-telegram-bot==13.15
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
six==1.16.0
soupsieve==2.3.2.post1
tornado==6.1
tzdata==2022.7
tzlocal==4.2
